+++ src/react/loading/loading.provider.test.unit.tsx
@@ -24,6 +24,22 @@ function TestConsumer() {
     );
 }
 
+/**
+ *
+ */
+function ContextCapture({ onCapture }: { onCapture: (val: any) => void }) {
+    onCapture(useLoading());
+    return null;
+}
+
+/**
+ *
+ */
+function Reader({ onRead }: { onRead: (val: unknown) => void }) {
+    onRead(React.use(LoadingContext));
+    return null;
+}
+
 describe('LoadingProvider', () => {
     it('should render children when not loading', () => {
         render(<LoadingProvider><span>content</span></LoadingProvider>);
@@ -60,14 +76,15 @@ describe('LoadingProvider', () => {
         /**
          *
          */
-        function ContextCapture() {
-            capturedContext = useLoading();
-            return null;
-        }
+        // ContextCapture component moved to module scope
+
+        const handleCapture = (val: any) => {
+            capturedContext = val;
+        };
 
         render(
             <LoadingProvider>
-                <ContextCapture />
+                <ContextCapture onCapture={handleCapture} />
             </LoadingProvider>,
         );
 
@@ -107,14 +124,10 @@ describe('useLoading', () => {
 describe('LoadingContext', () => {
     it('should default to undefined', () => {
         let contextValue: unknown;
-        /**
-         *
-         */
-        function Reader() {
-            contextValue = React.use(LoadingContext);
-            return null;
-        }
-        render(<Reader />);
+        const handleRead = (v: unknown) => {
+            contextValue = v;
+        };
+        render(<Reader onRead={handleRead} />);
         expect(contextValue).toBeUndefined();
     });
 });